Correction of the bug I could correct
