+++ app/src/main/res/layout/activity_mission_in_progress.xml
@@ -12,7 +12,7 @@
     tools:context=".MissionInProgressActivity">
 
     <com.google.android.gms.maps.MapView
-        android:id="@+id/mapInMissionView"
+        android:id="@+id/map_in_mission_view"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         app:layout_constraintVertical_bias="0.0" />
